move hooks and provider path to SRC
