--- src/components/AppContext.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const AppContext = React.createContext();
-
-export default AppContext;
